Ensure that isStrictDmd is False for Absent (fixes Trac #7737)
The demand <HyperStr, Absent> for a let-bound value is bit
strange; it means that the context will diverge, but this
argument isn't used. We don't want to use call-by-value here,
even though it's semantically sound if all bottoms mean
the same.

The fix is easy; just make "isStrictDmd" a bit more perspicuous.
See Note [Strict demands] in Demand.lhs