Adapt to tcRnGetInfo returning family instances too
This API change was part of the fix to Trac haskell/haddock#4175.  But it offers new
information to Haddock: the type-family instances, as well as the
class instances, of this type.

This patch just drops the new information on the floor, but there's an
open opportunity to use it in the information that Haddock displays.