This changes fixes a bad error in canonicalisation, concerning kind equality
We care careful not to construct a canonical equality whose LHS and RHS
have incompatible kinds.  (This is one of the invariants of a canonical
equality.)  See Note [Equalities with incompatible kinds].  However,
what I had not dealt with is when LHS and RHS *look* as if they have
different kinds, but after zonking they become the same.  Bad!

(This led to an ASSERT failure in the test for Trac #7176.)